WorkProcess/ExpertLevel/XUJINGYUAN/FacialExpressionClassifier/requirements.txt
timm>=0.6.0
swanlab>=0.2.0
scikit-learn>=1.0.0
numpy==2.2.6
Pillow>=8.0.0
tqdm>=4.60.0
matplotlib>=3.5.0
opencv-python==4.12.0.88
opencv-contrib-python==4.12.0.88
mediapipe>=0.10.0